pyproject.toml
[tool.poetry]
name = "fastapi-tutor"
version = "0.1.0"
description = "Tutorial de fastapi com pecistencoiia de dados deploy"
authors = ["Junior <eecfredes@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.0"
uvicorn = "^0.23.2"
psycopg = "^3.1.12"
psycopg-binary = "^3.1.12"
psycopg2-binary = "^2.9.9"
sqlalchemy = "^2.0.23"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
